refactor: :recycle: refactor common components export style
