--- src/components/common/layouts/MainLayout/ActivityBar/ActionItem/index.ts
@@ -1,3 +1,0 @@
-import ActionItem from "./ActionItem";
-
-export default ActionItem;
