+++ src/components/common/layouts/MainLayout/ActivityBar/index.style.ts
