--- src/components/common/layouts/MainLayout/ActivityBar/index.ts
@@ -1,3 +1,0 @@
-import ActivityBar from "./ActivityBar";
-
-export default ActivityBar;
