--- src/components/common/layouts/MainLayout/SideBar/index.ts
@@ -1,3 +1,0 @@
-import SideBar from "./SideBar";
-
-export default SideBar;
