+++ src/components/common/layouts/MainLayout/index.style.ts
