--- src/components/common/layouts/MainLayout/index.ts
@@ -1,3 +1,0 @@
-import MainLayout from "./MainLayout";
-
-export default MainLayout;
