--- src/components/common/molecules/AccordionSection/index.ts
@@ -1,3 +1,0 @@
-import AccordionSection from "./AccordionSection";
-
-export default AccordionSection;
