--- src/components/common/molecules/ThemeSwitcher/index.ts
@@ -1,3 +1,0 @@
-import ThemeSwitcher from "./ThemeSwitcher";
-
-export default ThemeSwitcher;
