+++ src/components/common/organisms/FileTree/FileTreeItem/index.style.ts
@@ -1,4 +1,4 @@
-import styled, { css } from "styled-components";
+import styled from "styled-components";
 
 const Wrapper = styled.div`
   display: flex;
