--- src/components/common/organisms/FileTree/FileTreeItem/index.ts
@@ -1,3 +1,0 @@
-import FileTree from "./FileTreeItem";
-
-export default FileTree;
