--- src/components/common/organisms/FileTree/index.ts
@@ -1,3 +1,0 @@
-import FileTree from "./FileTree";
-
-export default FileTree;
